requirements.txt
discord.py
openai.py
python-dotenv.py